+++ Plugins/Games/UNO/PlayerSettings.cs
@@ -1,13 +1,24 @@
 ﻿namespace UNO {
     public class PlayerSettings {
-        public AutoSortOptions AutoSort = AutoSortOptions.ByColour;
-        public HighlightOptions Highlight = HighlightOptions.Off;
-        public bool AllowDuelWithBot = true;
+        public AutoSortOptions AutoSort { get; set; }
+        public HighlightOptions Highlight { get; set; }
+        public bool AllowDuelWithBot { get; set; }
+        public bool Hints { get; set; }
+        public bool[] HintsSeen { get; private set; }
+
+        public PlayerSettings() {
+            this.AutoSort = AutoSortOptions.ByColour;
+            this.Highlight = HighlightOptions.Off;
+            this.AllowDuelWithBot = true;
+            this.Hints = true;
+            this.HintsSeen = new bool[UNOPlugin.Hints.Length];
+        }
 
         public bool IsDefault() {
             return (this.AutoSort == AutoSortOptions.ByColour &&
                     this.Highlight == HighlightOptions.Off &&
-                    this.AllowDuelWithBot);
+                    this.AllowDuelWithBot &&
+                    this.Hints);
         }
     }